+++ src/str.h
@@ -1,7 +1,7 @@
 #ifndef __STR_H
 #define __STR_H
 
-#include <stdint.h>
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -10,7 +10,7 @@ extern "C" {
 typedef struct str str_t;
 
 str_t *str_new(const char *s);
-str_t *str_new_len(const void *buf, uint32_t len);
+str_t *str_new_len(const void *buf, size_t len);
 void str_free(str_t *self);
 
 const char *sget(str_t *self);
