+++ src/vec.h
@@ -1,7 +1,7 @@
 #ifndef __VEC_H
 #define __VEC_H
 
-#include <stdint.h>
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -21,11 +21,11 @@ enum vec_alloc_type {
     //VEC_ALLOC_RANDOM,
 };
 
-vec_t *vec_new(uint32_t type_size, uint32_t cap);
-vec_t *vec_new_alloc(uint32_t type_size, uint32_t cap, enum vec_alloc_type alloc);
+vec_t *vec_new(size_t type_size, size_t cap);
+vec_t *vec_new_alloc(size_t type_size, size_t cap, enum vec_alloc_type alloc);
 void vec_free(vec_t *self);
 
-void *vat(vec_t *self, uint32_t idx);
+void *vat(vec_t *self, size_t idx);
 
 void vpush(vec_t *self, const void *value);
 void vpop(vec_t *self, /* out */ void *value);
@@ -33,21 +33,21 @@ void vpop(vec_t *self, /* out */ void *value);
 //void vpush_front(vec_t *self, const void *value);
 void vpop_back(vec_t *self, /* out */ void *value);
 
-void vpack(vec_t *self, const void *value, uint32_t cnt);
-void vdump(vec_t *self, /* out */ void *value, uint32_t cnt);
-void vdrop(vec_t *self, uint32_t cnt);
+void vpack(vec_t *self, const void *value, size_t cnt);
+void vdump(vec_t *self, /* out */ void *value, size_t cnt);
+void vdrop(vec_t *self, size_t cnt);
 void vshrink(vec_t *self);
-void vinsert(vec_t *self, uint32_t idx, const void *value, uint32_t cnt);
-void vremove(vec_t *self, uint32_t idx, uint32_t cnt);
+void vinsert(vec_t *self, size_t idx, const void *value, size_t cnt);
+void vremove(vec_t *self, size_t idx, size_t cnt);
 
 /**
  * vraw: only available on VEC_ALLOC_LINEAR
  */
 void *vraw(vec_t *self);
 
-uint32_t vtype(vec_t *self);
-uint32_t vsize(vec_t *self);
-uint32_t vcap(vec_t *self);
+size_t vtype(vec_t *self);
+size_t vsize(vec_t *self);
+size_t vcap(vec_t *self);
 
 #ifdef __cplusplus
 }
